[app-office/libreoffice-l10n-*] cleanup and fixing eclass

+++ eclass/libreoffice-l10n-2.eclass
@@ -40,38 +40,24 @@ L10N_VER="$(get_version_component_range 1-3)"
 L10N_RC_VERSION="$(get_version_component_range 4)"
 LO_BRANCH=$(get_version_component_range 1-2)
 
-if [ -n "${L10N_RC_VERSION}" ]; then
-	# this is a RC, thus testing
-	BASE_SRC_URI="http://download.documentfoundation.org/libreoffice/testing/${L10N_VER}/rpm"
-	TARBALL_VERSION="${L10N_VER}.${L10N_RC_VERSION}"
-else
-	BASE_SRC_URI="http://download.documentfoundation.org/libreoffice/stable/${L10N_VER}/rpm"
-	TARBALL_VERSION="${L10N_VER}"
-fi
-SRC_URI=""
-if [ "$(get_version_component_range 1)" = "3" ]; then
-	URI_PREFIX="LibO"
-	RPM_SUFFIX_LANG="langpack-rpm"
-	RPM_SUFFIX_HELP="helppack-rpm"
-else
-	URI_PREFIX="LibreOffice"
-	RPM_SUFFIX_LANG="rpm_langpack"
-	RPM_SUFFIX_HELP="rpm_helppack"
-fi
+BASE_SRC_URI="http://download.documentfoundation.org/${PN/-l10n/}/stable/${MY_PV}/rpm"
 
-# remove "name_part" when not needed
-if [[ ${PV} = 4.2.6.* || ${PV} = 4.2.6 ]]; then
-	name_part=-secfix
-else
-	name_part=
-fi
+SRC_URI=""
 
 # try guessing
 if [ "${LANGPACK_AVAIL}" = "1" ]; then
-	SRC_URI+="${BASE_SRC_URI}/x86/${URI_PREFIX}_${TARBALL_VERSION}${name_part}_Linux_x86_${RPM_SUFFIX_LANG}_${MY_LANG}.tar.gz"
+	langpack=""
+	[[ ${MY_LANG} == en ]] \
+		|| langpack="${BASE_SRC_URI}/x86/LibreOffice_${MY_PV}_Linux_x86_rpm_langpack_${MY_LANG/_/-}.tar.gz"
+	[[ -z ${langpack} ]] || SRC_URI+=" linguas_${MY_LANG}? ( ${langpack} )"
+	IUSE+=" linguas_${MY_LANG}"
 fi
+
 if [ "${HELPPACK_AVAIL}" = "1" ]; then
-	SRC_URI+=" ${BASE_SRC_URI}/x86/${URI_PREFIX}_${TARBALL_VERSION}${name_part}_Linux_x86_${RPM_SUFFIX_HELP}_${MY_LANG}.tar.gz"
+	helppack=""
+	[[ ${MY_LANG} == en ]] && lang2=${MY_LANG/en/en_US} || lang2=${MY_LANG}
+	helppack="offlinehelp? ( ${BASE_SRC_URI}/x86/LibreOffice_${MY_PV}_Linux_x86_rpm_helppack_${lang2/_/-}.tar.gz )"
+	SRC_URI+=" linguas_${MY_LANG}? ( ${helppack} )"
 fi
 
 unset lang helppack langpack lang2
@@ -89,16 +75,21 @@ libreoffice-l10n-2_src_unpack() {
 	local dir=${lang/_/-}
 	# for english we provide just helppack, as translation is always there
 	if [[ "${LANGPACK_AVAIL}" == "1" ]]; then
-		rpmdir="${URI_PREFIX}_${TARBALL_VERSION}"*"_Linux_x86_${RPM_SUFFIX_LANG}_${dir}/RPMS/"
-		# First remove dictionaries, we want to use system ones.
-		rm -rf "${S}/${rpmdir}/"*dict*.rpm
-		einfo "Unpacking Langpack"
-		rpm_unpack ./${rpmdir}/*.rpm
+		if [[ ${MY_LANG} != en ]]; then
+			rpmdir="LibreOffice_${PV}_Linux_x86_rpm_langpack_${dir}/RPMS/"
+			[[ -d ${rpmdir} ]] || die "Missing directory: \"${rpmdir}\""
+			# First remove dictionaries, we want to use system ones.
+			rm -rf "${S}/${rpmdir}/"*dict*.rpm
+			rpm_unpack "./${rpmdir}/"*.rpm
+		fi
 	fi
 	if [[ "${HELPPACK_AVAIL}" == "1" ]]; then
-		rpmdir="${URI_PREFIX}_${TARBALL_VERSION}"*"_Linux_x86_${RPM_SUFFIX_HELP}_${dir}/RPMS/"
-		einfo "Unpacking Helppack"
-		rpm_unpack ./${rpmdir}/*.rpm
+		if [[ "${LANGUAGES_HELP}" =~ " ${MY_LANG} " ]]; then
+			[[ ${MY_LANG} == en ]] && dir="en-US"
+			rpmdir="LibreOffice_${PV}_Linux_x86_rpm_helppack_${dir}/RPMS/"
+			[[ -d ${rpmdir} ]] || die "Missing directory: \"${rpmdir}\""
+			rpm_unpack ./"${rpmdir}/"*.rpm
+		fi
 	fi
 }
 
@@ -107,13 +98,13 @@ libreoffice-l10n-2_src_configure() { :; }
 libreoffice-l10n-2_src_compile() { :; }
 
 libreoffice-l10n-2_src_install() {
-	local dir="${S}"/opt/${PN/-l10n/}$(get_version_component_range 1-2)/
-	# Condition required for people that do not install anything eg no linguas
-	# or just english with no offlinehelp.
-	if [[ -d "${dir}" ]] ; then
-		insinto /usr/$(get_libdir)/${PN/-l10n/}/
-		doins -r "${dir}"/*
-	fi
-	# remove extensions that are in the l10n for some weird reason
-	rm -rf "${ED}"/usr/$(get_libdir)/${PN/-l10n/}/share/extensions/
+local dir="${S}"/opt/libreoffice${LO_BRANCH}/
+# Condition required for people that do not install anything eg no linguas
+# or just english with no offlinehelp.
+if [[ -d "${dir}" ]] ; then
+	insinto /usr/$(get_libdir)/libreoffice/
+	doins -r "${dir}"/*
+fi
+# remove extensions that are in the l10n for some weird reason
+rm -rf "${ED}"/usr/$(get_libdir)/libreoffice/share/extensions/
 }
